Draft - Create/Update recommendation

+++ src/main/java/school/faang/user_service/service/RecommendationService.java
@@ -2,6 +2,7 @@ package school.faang.user_service.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import school.faang.user_service.dto.recommendation.RecommendationDto;
 import school.faang.user_service.dto.recommendation.SkillOfferDto;
 import school.faang.user_service.entity.Skill;
@@ -17,8 +18,10 @@ import school.faang.user_service.repository.UserSkillGuaranteeRepository;
 import school.faang.user_service.repository.recommendation.RecommendationRepository;
 import school.faang.user_service.repository.recommendation.SkillOfferRepository;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -31,67 +34,59 @@ public class RecommendationService {
     private final RecommendationMapper recommendationMapper;
     private final UserSkillGuaranteeRepository userSkillGuaranteeRepository;
 
+    @Transactional
     public RecommendationDto create(RecommendationDto recommendationDto) {
-        Recommendation recommendation = saveNewRecommendation(recommendationDto);
-        saveNewSkillOffers(recommendationDto);
+        Long id = recommendationRepository.create(recommendationDto.getAuthorId(),
+                recommendationDto.getReceiverId(),
+                recommendationDto.getContent());
+        Recommendation recommendation  = recommendationRepository.findById(id).orElseThrow();
+        recommendation.setCreatedAt(LocalDateTime.now());
+        recommendation.setSkillOffers(initNewSkillOffers(recommendation, recommendationDto.getSkillOffers()));
         return recommendationMapper.toDto(recommendationRepository.save(recommendation));
-    }
-
-    private Recommendation saveNewRecommendation(RecommendationDto recommendationDto) {
-        Recommendation recommendation = recommendationMapper.toEntity(recommendationDto);
-        User author = userRepository.findById(recommendationDto.getAuthorId()).orElseThrow();
-        User receiver = userRepository.findById(recommendationDto.getReceiverId()).orElseThrow();
-        recommendation.setAuthor(author);
-        recommendation.setReceiver(receiver);
-        recommendation.setContent(recommendationDto.getContent());
-        recommendation.setCreatedAt(recommendationDto.getCreatedAt());
-        return recommendation;
-    }
-
-    private Recommendation saveNewSkillOffers(RecommendationDto recommendationDto) {
-        Recommendation recommendation = recommendationMapper.toEntity(recommendationDto);
-        List<SkillOfferDto> skillOfferDtos = recommendationDto.getSkillOffers();
-        validate(skillOfferDtos);
-        List<Skill> offeredSkills = skillRepository.findAllById(skillOfferDtos.stream().map(SkillOfferDto::getSkillId).collect(Collectors.toList()));
-        addGuarantee(recommendation, offeredSkills);
-        recommendation.setSkillOffers(offeredSkills.stream().map(skill -> SkillOffer.builder().skill(skill).build()).collect(Collectors.toList()));
-        return recommendation;
     }
 
     public RecommendationDto update(RecommendationDto updated) {
-    Recommendation recommendation = recommendationRepository.findById(updated.getId()).orElseThrow();
-        recommendationRepository.update(
-                updated.getAuthorId(),
-                updated.getReceiverId(),
-                updated.getContent()
-        );
-
+        if(updated.getId() == null){
+            throw new DataValidationException("entity not found exception");
+        }
+        Recommendation recommendation = recommendationMapper.toEntity(updated);
         skillOfferRepository.deleteAllByRecommendationId(recommendation.getId());
-        saveNewSkillOffers(updated);
+        recommendation.setSkillOffers(initNewSkillOffers(recommendation, updated.getSkillOffers()));
+        recommendation.setUpdatedAt(updated.getCreatedAt() !=null ? updated.getCreatedAt() : LocalDateTime.now());
         return recommendationMapper.toDto(recommendationRepository.save(recommendation));
+    }
+
+    private List<SkillOffer> initNewSkillOffers(Recommendation recommendation, List<SkillOfferDto> skillOffers) {
+        validate(skillOffers);
+        List<Skill> offeredSkills = skillRepository.findAllById(
+                skillOffers.stream()
+                .map(SkillOfferDto::getSkillId)
+                .collect(Collectors.toList()));
+        addGuarantee(recommendation, offeredSkills);
+        return  skillOfferRepository.findAllById(
+                offeredSkills.stream()
+                        .map(s -> skillOfferRepository.create(s.getId(), recommendation.getId()))
+                        .collect(Collectors.toList()));
     }
 
     private void addGuarantee(Recommendation recommendation, List<Skill> offeredSkills) {
         List<Skill> receiverSkills = recommendation.getReceiver().getSkills();
         List<UserSkillGuarantee> userSkillGuarantees = new ArrayList<>();
-        offeredSkills.forEach(offeredSkill -> {
-                    receiverSkills.forEach(receiverSkill -> {
-                        boolean userSkillContainGuarantor = !receiverSkill.getGuarantees()
+        offeredSkills.forEach(
+                s -> {
+                    if (receiverSkills.contains(s)) {
+                        if (s.getGuarantees()
                                 .stream()
                                 .map(UserSkillGuarantee::getGuarantor)
-                                .toList()
-                                .contains(recommendation.getAuthor());
-                        UserSkillGuarantee userSkillGuarantee = UserSkillGuarantee.builder()
-                                .skill(offeredSkill)
-                                .guarantor(recommendation.getAuthor())
-                                .user(recommendation.getReceiver())
-                                .build();
-                        if (receiverSkills.contains(offeredSkill)) {
-                            if (!userSkillContainGuarantor) {
-                                userSkillGuarantees.add(userSkillGuarantee);
-                            }
+                                .noneMatch(g -> g.equals(recommendation.getAuthor()))) {
+                            UserSkillGuarantee userSkillGuarantee = UserSkillGuarantee.builder()
+                                    .skill(s)
+                                    .guarantor(recommendation.getAuthor())
+                                    .user(recommendation.getReceiver())
+                                    .build();
+                            userSkillGuarantees.add(userSkillGuarantee);
                         }
-                    });
+                    }
                 }
         );
         userSkillGuaranteeRepository.saveAll(userSkillGuarantees);
@@ -101,8 +96,10 @@ public class RecommendationService {
         if (skills != null && !skills.isEmpty()) {
             List<Long> skillIds = skills.stream()
                     .map(SkillOfferDto::getSkillId)
+                    .filter(Objects::nonNull)
+                    .distinct()
                     .collect(Collectors.toList());
-            if (!skillRepository.existsAllById(skillIds)) {
+            if (skillRepository.countExisting(skillIds) != skills.size()) {
                 throw new DataValidationException("list of skills contains not valid skills, please, check this");
             }
         }
